Modified "Getting widget client object" example
I found the "Getting widget client object" to confusing with the Id of
the control being "grid" the variable the result of the selctor is being
assigned to named "grid" and then having the "kendoGrid" as the
parameter in the data method.  Maybe it was just me, but it took me a
bit to figure it out. ;)